refactor: metadata search (#1)
